- fix typo  - remove unused import  - fix unchecked warning (+1 squashed commits)
