+++ commons-server/src/main/java/com/navercorp/pinpoint/common/server/bo/serializer/trace/v1/TraceRowKeyEncoderV1.java
@@ -1,7 +1,6 @@
 package com.navercorp.pinpoint.common.server.bo.serializer.trace.v1;
 
 import com.navercorp.pinpoint.common.PinpointConstants;
-import com.navercorp.pinpoint.common.server.bo.BasicSpan;
 import com.navercorp.pinpoint.common.server.bo.serializer.RowKeyEncoder;
 import com.navercorp.pinpoint.common.util.BytesUtils;
 import com.navercorp.pinpoint.common.util.TransactionId;
