+++ plugins/spring/src/main/java/com/navercorp/pinpoint/plugin/spring/beans/interceptor/TargetBeanFilter.java
@@ -20,7 +20,6 @@ import java.lang.annotation.Annotation;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.regex.Pattern;
 
 import com.navercorp.pinpoint.bootstrap.config.ProfilerConfig;
 import com.navercorp.pinpoint.bootstrap.util.PathMatcher;
