+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/instrument/ASMAspectWeaver.java
@@ -139,7 +139,7 @@ public class ASMAspectWeaver {
         }
     }
 
-    private class MethodNodes {
+    private static class MethodNodes {
         public final List<ASMMethodNodeAdapter> pointCuts = new ArrayList<ASMMethodNodeAdapter>();
         public final List<ASMMethodNodeAdapter> jointPoints = new ArrayList<ASMMethodNodeAdapter>();
         public final List<ASMMethodNodeAdapter> utils = new ArrayList<ASMMethodNodeAdapter>();
