+++ profiler/src/test/java/com/navercorp/pinpoint/profiler/instrument/ASMClassNodeAdapterTest.java
@@ -120,7 +120,7 @@ public class ASMClassNodeAdapterTest {
                 adapter.addGetterMethod(getterMethodName, field);
             }
         });
-        Class clazz = classLoader.loadClass(targetClassName);
+        Class<?> clazz = classLoader.loadClass(targetClassName);
         Method method = clazz.getDeclaredMethod(getterMethodName);
         assertEquals(0, method.invoke(clazz.newInstance()));
     }
@@ -140,7 +140,7 @@ public class ASMClassNodeAdapterTest {
                 adapter.addSetterMethod(setterMethodName, field);
             }
         });
-        Class clazz = classLoader.loadClass(targetClassName);
+        Class<?> clazz = classLoader.loadClass(targetClassName);
         Method method = clazz.getDeclaredMethod(setterMethodName, int.class);
         method.invoke(clazz.newInstance(), 10);
     }
