+++ web/src/main/java/com/navercorp/pinpoint/web/controller/ConfigController.java
@@ -63,7 +63,7 @@ public class ConfigController {
         }
         
         if(!StringUtils.isEmpty(webProperties.getSecurityGuideUrl())) {
-        	result.put("securityGuideUrl", webProperties.getSecurityGuideUrl());
+            result.put("securityGuideUrl", webProperties.getSecurityGuideUrl());
         }
         
         return result;
