+++ web/src/main/java/com/navercorp/pinpoint/web/dao/TraceDao.java
@@ -17,7 +17,6 @@
 package com.navercorp.pinpoint.web.dao;
 
 
-import java.util.Collection;
 import java.util.List;
 
 import com.navercorp.pinpoint.common.server.bo.SpanBo;
