+++ web/src/main/java/com/navercorp/pinpoint/web/filter/LinkFilter.java
@@ -112,11 +112,11 @@ public class LinkFilter implements Filter {
 
         this.rpcHintList = this.filterHint.getRpcHintList(fromApplicationName);
         // TODO fix : fromSpan base rpccall filter
-        this.acceptURLFilter = cratePatternFilter(filterDescriptor);
+        this.acceptURLFilter = createPatternFilter(filterDescriptor);
         logger.info("acceptURLFilter:{}", acceptURLFilter);
     }
 
-    private URLPatternFilter cratePatternFilter(FilterDescriptor filterDescriptor) {
+    private URLPatternFilter createPatternFilter(FilterDescriptor filterDescriptor) {
         if (filterDescriptor.getUrlPattern() == null) {
             return new BypassURLPatternFilter();
         }
